refactor: move from exceptions to exception
