--- src/middleware/auth.middleware.ts
@@ -1,4 +1,0 @@
-import {} from "express"
-import jwt from "jsonwebtoken"
-
-
