+++ src/middleware/error.middleware.ts
@@ -1,5 +1,5 @@
 import { NextFunction, Request, Response } from 'express'
-import HttpException from '../exceptions/HttpException'
+import HttpException from '../exception/HttpException'
 import endpoints from '../utils/endpoints'
 
 function errorMiddleware(
