+++ src/middleware/validation.middleware.ts
@@ -2,7 +2,7 @@ import 'reflect-metadata' // required for class-transformer
 import { plainToInstance } from 'class-transformer'
 import { validate, ValidationError } from 'class-validator'
 import { RequestHandler } from 'express'
-import HttpException from '../exceptions/HttpException'
+import HttpException from '../exception/HttpException'
 
 function validationMiddleware(type: any): RequestHandler {
   return (req, _, next) => {
