+++ src/post/post.controller.ts
@@ -3,8 +3,8 @@ import IPost from './post.interface'
 import IController from '../interface/controller.interface'
 import PostModel from './post.model'
 import { Types } from 'mongoose'
-import HttpException from '../exceptions/HttpException'
-import PostNotFoundException from '../exceptions/PostNotFoundException'
+import HttpException from '../exception/HttpException'
+import PostNotFoundException from '../exception/PostNotFoundException'
 import validationMiddleware from '../middleware/validation.middleware'
 import CreatePostDto from './post.dto'
